Move remaining managers
